finalized basic blueprint
