+++ genetic_algorithm/environment.py
@@ -1,0 +1,5 @@
+class Individual:
+    pass
+
+class Population(Individual):
+    pass
